遇到问题：上传太大占用浏览器内存，在send的时候才会出现，而在文件引用的时候不占很多内存

+++ Models/FileInfo.cs
@@ -30,10 +30,10 @@ namespace WebAppMVC.Models
                 return this.FileName.Substring(this.FileName.LastIndexOf('.'));
             }
         }
-        public int FileSize { get; set; }
+        public long FileSize { get; set; }
         public string SavePath { get; set; }
 
         public List<byte> FileData { get; set; }
-        public int CurrentSize { get; set; }
+        public long CurrentSize { get; set; }
     }
 }
